% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loggr_redis.R
\name{format_log_entry_json}
\alias{format_log_entry_json}
\title{JSON log entry formatter.}
\usage{
format_log_entry_json(event)
}
\arguments{
\item{event}{The log event to format}
}
\description{
JSON log entry formatter.
}

