% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loggr_redis.R
\name{log_redis}
\alias{log_redis}
\title{Activate logging to Redis}
\usage{
log_redis(con, key, ..., .warning = TRUE, .error = TRUE, .message = TRUE,
  .formatter = format_log_entry_json, subscriptions = NULL)
}
\arguments{
\item{con}{A redis connection object, created by
\code{RedisAPI::hiredis}.}

\item{key}{Key to log to.  If this exists it \emph{must} be a
list (this is enforced in this function but must remain true while
logging is active).}

\item{...}{list of quoted or unquoted events to log. In none are
provided all log events will be captured.}

\item{.warning}{logical: capture regular warnings (\code{simpleWarning})?}

\item{.error}{logical: capture regular errors (\code{simpleError})?}

\item{.message}{logical: capture regular messages (\code{simpleMessage})?}

\item{.formatter}{function: the formatting function used to convert
a log event to its character representation for the log file.
Note that the default formatter here is to JSON, not to text.}

\item{subscriptions}{character vector: optional list of
subscriptions to use (in place of specifying with \code{...}).}
}
\description{
Activate logging to Redis
}

